Revert "Move blocks commands under mimirtool blocks; add as Job to Jsonnet (#…"
This reverts commit e0f34e9eb401d6eab1719e502be54c0cbc5c9b1d.
