+++ operations/mimir/images.libsonnet
@@ -20,6 +20,5 @@
 
     query_tee: 'grafana/query-tee:3.1.2',
     continuous_test: self.mimir,
-    mimirtool: 'grafana/mimirtool:3.1.1',
   },
 }
