+++ tools/mark-blocks/main_test.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package mark
+package main
 
 import (
 	"bytes"
